LayoutTests/imported/w3c/web-platform-tests/custom-elements/CustomElementRegistry-getName-expected.txt


PASS customElements.getName must return null when the registry does not contain an entry with the given constructor
FAIL customElements.getName must throw when the element interface is not a constructor assert_throws_js: customElements.getName must throw a TypeError when the element interface is undefined function "function () { customElements.getName(undefined); }" did not throw
FAIL customElements.getName returns the name of the entry with the given constructor when there is a matching entry. assert_throws_js: customElements.getName must throw a TypeError when the element interface is undefined function "function () { customElements.getName(customElements.getName(OtherExistingCustomElement)); }" did not throw
PASS customElements.getName returns the name of the entry with the given customized built in constructor when there is a matching entry.
